processor/requirements.txt
kafka-python
psycopg2-binary
PyYAML
python-dotenv